+++ slang.c
@@ -35,7 +35,7 @@ int main(int argc, char **argv) {
         printAllOscillators(main_interpreter);
 
         SlangBufferCore* sbc = createBufferCore(main_interpreter, 48000, 512);
-        double* buf = renderBuffer(sbc);
+        float* buf = renderBuffer(sbc);
         printAudioBuffer(buf, 512);
         free(buf);
